Set the PLL sampling rate to int

+++ include/ecss/pll.h
@@ -59,7 +59,7 @@ namespace gr {
         * \param freq_central central value of frequency that PLL can catch.
         * \param bw bandwidth of frequency that PLL can catchl.
        */
-      static sptr make(float samp_rate, int N, const std::vector<double> &coefficients, float freq_central, float bw);
+      static sptr make(int samp_rate, int N, const std::vector<double> &coefficients, float freq_central, float bw);
 
 
       /*******************************************************************
