+++ lib/pll_impl.h
@@ -30,7 +30,7 @@ namespace gr {
     class pll_impl : public pll
     {
       int d_N;
-      float d_samp_rate;
+      int d_samp_rate;
       float d_bw;
       int64_t d_integer_phase;
       double d_integer_phase_denormalized;                /*!< Integer value after to be denormalized */
@@ -99,7 +99,7 @@ namespace gr {
       double frequency_limit(double step);
 
       public:
-        pll_impl(float samp_rate, int N, const std::vector<double> &coefficients, float freq_central, float bw);
+        pll_impl(int samp_rate, int N, const std::vector<double> &coefficients, float freq_central, float bw);
         ~pll_impl();
 
         int work(int noutput_items,
